Remaining problems:
- GridObject inherits IRenderable and IPosition rather than being a separate entity, so it cannot be added to the renderer directly. Maybe have a list of permanent RendObjs in the renderer?
- Assimp giving NOMINMAX problems

+++ src/EngineGUI/AnimationSequencer.cpp
@@ -12,7 +12,7 @@ AnimationSequencer::AnimationSequencer() :mFrameMin(0), mFrameMax(0) {};
 
 int AnimationSequencer::GetItemTypeCount() const 
 {
-    return scene->allNodes.size();
+    return (i32)scene->allNodes.size();
 }
 
 const char* AnimationSequencer::GetItemTypeName(int index) const
